Translated .f.frag files to Fortran 2003 bindings

+++ EXAMPLE_LEGOS/test_legos/NEIGH_RVEC_update_Rij_vectors.f.frag
@@ -5,19 +5,20 @@
 !-------------------------------------------------------------------------------
 subroutine NEIGH_RVEC_update_Rij_vectors(DIM, N, coords, NNMAX, &
                                          neighborList, RijList)
+  use, intrinsic :: iso_c_binding
   implicit none
 
   !-- Transferred variables
-  integer,            intent(in)  :: DIM
-  integer,            intent(in)  :: N
-  double precision,   intent(in)  :: coords(DIM,N)
-  integer,            intent(in)  :: NNMAX
-  integer,            intent(in)  :: neighborList(NNMAX+1,N)
-  double precision,   intent(out) :: RijList(DIM,NNMAX+1,N)
+  integer(c_int), intent(in)  :: DIM
+  integer(c_int), intent(in)  :: N
+  real(c_double), intent(in)  :: coords(DIM,N)
+  integer(c_int), intent(in)  :: NNMAX
+  integer(c_int), intent(in)  :: neighborList(NNMAX+1,N)
+  real(c_double), intent(out) :: RijList(DIM,NNMAX+1,N)
 
   !-- Local variables
-  integer i, j, jj, NN
-  double precision dx(DIM)
+  integer(c_int) i, j, jj, NN
+  real(c_double) dx(DIM)
 
   do i=1,N
      NN = neighborList(1,i)
